# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KDown Local API",
    "description": "Local KDown download manager service",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/roots": {
      "get": {
        "tags": [
          "kdown"
        ],
        "summary": "Lists configured roots with canonical paths.",
        "operationId": "list_roots",
        "responses": {
          "200": {
            "description": "Root list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RootDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "kdown"
        ],
        "summary": "Adds an existing directory as a configured root. This is the explicit\nauthorization step, so an absolute path is expected here.",
        "operationId": "create_root",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRootRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Root added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RootDto"
                }
              }
            }
          },
          "422": {
            "description": "Path is not an existing directory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/roots/{id}": {
      "get": {
        "tags": [
          "kdown"
        ],
        "summary": "One configured root.",
        "operationId": "get_root",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Root",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RootDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "kdown"
        ],
        "summary": "Updates label, enabled state, or default flag of a root.",
        "operationId": "patch_root",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRootRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Root updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RootDto"
                }
              }
            }
          },
          "409": {
            "description": "Root in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorEnvelope": {
        "type": "object",
        "description": "Machine-stable error envelope: `code` identifies the failure class,\n`message` is a safe user-facing summary, and optional fields carry\nstructured context. Secrets, URL query values, and credentials never\nappear in `message` or `detail`.",
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "current_job": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobViewDto"
              }
            ]
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "field_errors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
...
      "CreateRootRequest": {
        "type": "object",
        "description": "Request to add an existing directory as a configured root.",
        "required": [
          "label",
          "absolute_path"
        ],
        "properties": {
          "absolute_path": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "make_default": {
            "type": "boolean"
          }
        }
      },
      "DesiredStateDto": {
        "type": "string",
        "description": "User-facing desired state.",
        "enum": [
          "Running",
          "Paused",
          "Cancelled"
        ]
      },
      "DurableJobStatusDto": {
        "type": "string",
        "description": "Durable lifecycle status as the UI presents it.",
        "enum": [
          "Queued",
          "Recovering",
          "Active",
          "Paused",
          "Completed",
          "Failed",
          "Cancelled"
        ]
      },
      "EngineSnapshotViewDto": {
        "type": "object",
        "description": "Display-safe telemetry snapshot.",
        "required": [
          "state_label",
          "bytes_received",
          "network_bytes",
          "reused_bytes",
          "retries",
          "elapsed_ms"
        ],
        "properties": {
          "bytes_received": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "elapsed_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "network_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "retries": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "reused_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "state_label": {
            "type": "string"
          }
        }
      },
...
      "JobViewDto": {
        "type": "object",
        "description": "Display-safe job view. Source URLs are redacted; filesystem paths are\nreduced to relative display form.",
        "required": [
          "id",
          "status",
          "desired_state",
          "control_version",
          "sample_seq",
          "source_display",
          "root_id",
          "root_label",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "attempt_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "control_version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "desired_state": {
            "$ref": "#/components/schemas/DesiredStateDto"
          },
          "destination_display": {
            "type": [
              "string",
              "null"
            ],
            "description": "Final display destination relative to the root, once completed."
          },
          "filename_override": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "relative_directory": {
            "type": [
              "string",
              "null"
            ]
          },
          "root_id": {
            "type": "string",
            "format": "uuid"
          },
          "root_label": {
            "type": "string",
            "description": "Root label; ordinary job views never carry absolute paths."
          },
          "sample_seq": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "snapshot": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EngineSnapshotViewDto"
              }
            ]
          },
          "source_display": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DurableJobStatusDto"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PatchRootRequest": {
        "type": "object",
        "description": "Partial root update.",
        "properties": {
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "make_default": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
      "RootDto": {
        "type": "object",
        "description": "A configured root as the administration surface presents it. This is\nthe only DTO family that may carry absolute filesystem paths.",
        "required": [
          "id",
          "label",
          "canonical_path",
          "enabled",
          "is_default",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "canonical_path": {
            "type": "string"
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_default": {
            "type": "boolean"
          },
          "label": {
            "type": "string"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  },
...
}